Initial attempt to add Filters (for loggers and handlers).
I've opted for filters to take a record and return true/false on whether to write the record.
This choice has the benefit of relying on a relatively simple interface for both loggers and handlers, but
requires that a Record be created first. As a result, I'm choosing to leave the initial logger filtering by level untouched as I don't want to add the extra overhead yet (I may change this in the future)..